add a license file (mostly GPL 2+)
This LICENSE allows Turtle Crossing as a whole to be distributed under
the GNU GPLv2 or later, but includes a note that components may be
licensed differently. I intend for the OSNAP core and Turtle Crossing's
specific modifications to stay GPL, but reusable apps developed to
support them will probably be released under MIT/X11.