rework navbar and color scheme
The colors are now inspired by the official Django docs -- I wanted
something with dark green navigation, and figured that Django's look
rather nice.

The navbar is now fixed to the top of the page, but doesn't scroll with
the viewport. It also includes a notification of which user is logged in.
(Actual login/logout still runs through the admin, but Persona will be
added soon.)